Comment out Pipeline, StateAssembler, and Runner
