+++ lib/lightning_web/live/run_live/components.ex
@@ -1,5 +1,6 @@
 defmodule LightningWeb.RunLive.Components do
   @moduledoc false
+  alias Lightning.Invocation
   use LightningWeb, :component
   import LightningWeb.RouteHelpers
   alias Lightning.Pipeline
@@ -176,7 +177,8 @@ defmodule LightningWeb.RunLive.Components do
       assign(
         assigns,
         :log,
-        Pipeline.logs_for_run(assigns.run) |> Enum.map(fn log -> log.message end)
+        Invocation.logs_for_run(assigns.run)
+        |> Enum.map(fn log -> log.message end)
       )
 
     ~H"""
