+++ test/lightning/invocation_test.exs
@@ -193,7 +193,7 @@ defmodule Lightning.InvocationTest do
                )
 
       assert run.exit_code == 42
-      assert run |> Pipeline.logs_for_run() == []
+      assert run |> Invocation.logs_for_run() == []
       assert run.finished_at == ~U[2022-02-02 11:49:00.000000Z]
       assert run.started_at == ~U[2022-02-02 11:49:00.000000Z]
     end
@@ -225,7 +225,7 @@ defmodule Lightning.InvocationTest do
       assert {:ok, %Run{} = run} = Invocation.update_run(run, update_attrs)
       assert run.exit_code == 43
       assert run.finished_at == ~U[2022-02-03 11:49:00.000000Z]
-      assert Pipeline.logs_for_run(run) == []
+      assert Invocation.logs_for_run(run) == []
       assert run.started_at == ~U[2022-02-03 11:49:00.000000Z]
     end
