+++ test/lightning_web/end_to_end_test.exs
@@ -133,10 +133,10 @@ defmodule LightningWeb.EndToEndTest do
       assert run_1.finished_at != nil
       assert run_1.exit_code == 0
 
-      assert Pipeline.assemble_logs_for_run(run_1) =~ "Done in"
+      assert Invocation.assemble_logs_for_run(run_1) =~ "Done in"
 
       r1_logs =
-        Pipeline.logs_for_run(run_1)
+        Invocation.logs_for_run(run_1)
         |> Enum.map(fn line -> line.body end)
 
       # Check that versions are accurate and printed at the top of each run
@@ -156,14 +156,14 @@ defmodule LightningWeb.EndToEndTest do
       assert run_2.finished_at != nil
       assert run_2.exit_code == 1
 
-      log = Pipeline.assemble_logs_for_run(run_2)
+      log = Invocation.assemble_logs_for_run(run_2)
       assert log =~ ~S[{"password":"***","username":"quux"}]
       assert log =~ ~S[Error in runtime execution!]
 
       #  Run 3 should succeed and log "6"
       assert run_3.finished_at != nil
       assert run_3.exit_code == 0
-      log = Pipeline.assemble_logs_for_run(run_3)
+      log = Invocation.assemble_logs_for_run(run_3)
       assert log =~ "[JOB] ℹ 6"
     end)
   end
